demo/sam/pyproject.toml
[project]
name = "sam-demo"
version = "0.1.0"
description = "Real SAM inference demo using sam-prompter"
requires-python = ">=3.12"
dependencies = [
    "sam-prompter",
    "torch==2.9.1",
    "transformers>=4.39",
    "gradio>=6.6.0",
    "Pillow>=12.1",
    "numpy>=2.4",
    "spaces>=0.47.0",
]

[tool.uv.sources]
sam-prompter = { path = "../../", editable = true }

[tool.ruff]
extend = "../../pyproject.toml"

[tool.ruff.lint.per-file-ignores]
"*.py" = ["INP001"]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
